requirements.txt
awscli==1.27.41
beautifulsoup4==4.7.1
bleach==6.0.0
certifi==2019.3.9
chardet==3.0.4
click==8.0.3
docutils==0.14
exceptiongroup==1.1.1
Flask==2.0.2
idna==2.8
importlib-metadata
iniconfig==2.0.0
itsdangerous==2.0.1
jaraco.classes==3.2.3
Jinja2==3.0.2
jmespath==0.9.4
keyring==23.13.1
markdown-it-py==2.2.0
MarkupSafe==2.0.1
mdurl==0.1.2
mido==1.2.9
more-itertools==9.1.0
packaging==23.1
pkginfo==1.9.6
pluggy==1.0.0
pyasn1==0.4.8
PyAudio==0.2.13
pydub==0.23.1
Pygments==2.15.1
pytest==7.3.1
python-dateutil==2.8.0
python-magic==0.4.27
pytube==15.0.0
readme-renderer==37.3
requests
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.3.5
rsa==4.7.2
s3transfer
six==1.12.0
soupsieve==1.9
tomli==2.0.1
twine==4.0.2
typing-extensions==3.10.0.2
webencodings==0.5.1
Werkzeug==2.0.2
youtube-dl @ git+https://github.com/ytdl-org/youtube-dl.git@211cbfd5d46025a8e4d8f9f3d424aaada4698974
youtube-dlc==2020.11.11.post3
zipp==3.6.0
